Move DataMart generation in separate class #3
